+++ appglue/src/main/java/com/appglue/ShortcutActivity.java
@@ -1,12 +1,13 @@
 package com.appglue;
 
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+
 import com.appglue.engine.CompositeService;
 import com.appglue.engine.OrchestrationService;
 import com.appglue.serviceregistry.Registry;
 
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
 import static com.appglue.Constants.COMPOSITE_ID;
 import static com.appglue.Constants.DURATION;
 import static com.appglue.Constants.RUN_NOW;
@@ -21,7 +22,7 @@ public class ShortcutActivity extends Activity
 	        Bundle b = stuff.getExtras();
 	        
 	        // Maybe if it isn't set we could ask them which one they want to execute?
-	        long id = b.getLong(COMPOSITE_ID);
+	        int id = b.getInt(COMPOSITE_ID);
 	        
 	        Registry registry = Registry.getInstance(this);
 			CompositeService cs = registry.getComposite(id);
